1.添加预期上线枚举类 2.完成评论功能 3.修正预期上线时间类型 4.完成通知功能

+++ dal/src/main/java/com/timevale/forward/dal/entity/BizDemandDO.java
@@ -79,7 +79,7 @@ public class BizDemandDO extends BaseDO {
     /**
      * 计划发布日期
      */
-    Integer planReleaseDate;
+    Byte planReleaseDate;
 
     /**
      * 驳回理由
